Rewrite for FF126 support
*THIS UPDATES BREAKS SUPPORT FOR FF125 AND BELOW*

Changed a LOT of stuff with this.

`Important Note` theme config preferences have been changed with this version! Have a look below for the new ones
`Sidebery` Sidebery is now themed using a preference in about:config. Please delete the css in `sidebery settings` > `style editor` to make sure everything looks good

`Fixed`
- #30 Transucency now works on FF 126/127 'macOS'
- #32 Transparent window/navbar/urlbar bug in windows and linux fixed

`New`
- the minimal "arc" sidebar is now the default style.

- Changed theme preferences for about:config
the new ones are as follows
- `af.edgyarc.edge-sidebar` Makes the sidebar look more like edge (previous default view)
- `af.edgyarc.macos-translucent` enable translucent windows in macos
- `af.edgyarc.show-sidebar-header` Shows sidebar Header
- `af.edgyarc.thin-navbar` Thin navbar
   - `af.edgyarc.minimal-navbar` hide everything in the navbar apart from the url until hover
- `af.sidebery.edgyarc-theme` Sidebery theme for edgyarc
- `af.edgyarc.centered-url` - center urlbar text